It seems that we can read ndarrays now

+++ src/main/java/org/asdfformat/asdf/ndarray/LongNdArray.java
@@ -5,7 +5,7 @@ package org.asdfformat.asdf.ndarray;
  * long elements.
  */
 public interface LongNdArray extends NdArray<LongNdArray> {
-    long get(long... indices);
+    long get(int... indices);
 
     long[] toArray(long[] array);
